Introduction and Background
I attempted to describe the Introduction and Background of the part of speech tagging idea for the code-switching analysis.  Mandarin-English bilingual speakers' word retrieval rate is higher for verbs while their accuracy in word retrieval is higher for verbs. We expect retrieval rate to be a greater indicator of which non-native words occur in fluent mixed-language conversation, so non-native verbs should be more frequent than non-native nouns. Non-native words are in Mandarin for dataset1.txt  and in English for datset2.txt